Make sure job fails if error occurs when publishing to staging

+++ _scripts/publish-to-staging.sh
@@ -8,8 +8,23 @@ rake clean[all]
 
 # Make sure there was no update to the used dependencies (if not, this is just a quick version check for Bundler)
 rake setup
+rc=$?
+if [[ $rc != 0 ]] ; then
+    echo "ERROR: Rake setup failed!"
+    exit $rc
+fi
 
 # Build the site using the staging profile and sync
 rake --trace gen[staging]
+rc=$?
+if [[ $rc != 0 ]] ; then
+    echo "ERROR: Site generation failed!"
+    exit $rc
+fi
 
 rsync --delete -avh _site/ 54.174.65.136:/var/www/staging-beanvalidation.org
+rc=$?
+if [[ $rc != 0 ]] ; then
+    echo "ERROR: Site sync failed!"
+    exit $rc
+fi
